Upgrade dependencies

+++ src/datamodel.rs
@@ -32,7 +32,7 @@ pub(crate) fn get_index_model() -> Vec<bson::Document> {
     vec![doc! {
         "key": doc! { "t" : 1 },
         "name": "expiration",
-        "expireAfterSeconds": TIME_TO_LIVE.as_secs(),
+        "expireAfterSeconds": TIME_TO_LIVE.as_secs() as i64,
     }]
 }
